31 user keresés backend

+++ server/src/users/users.service.ts
@@ -1,7 +1,9 @@
 import {
   BadRequestException,
+  ForbiddenException,
   Injectable,
   Logger,
+  NotFoundException,
   PreconditionFailedException,
   UnauthorizedException,
 } from "@nestjs/common";
@@ -17,6 +19,7 @@ import { ChangePasswordDto } from "./dtos/change-password.dto";
 import { FinishRegistrationDto } from "./dtos/finish-registration.dto";
 import { LoginUserDto } from "./dtos/login-user.dto";
 import { RegisterUserDto } from "./dtos/register-user.dto";
+import { SearchUserDto } from "./dtos/search-user.dto";
 import { User } from "./entities/users.entity";
 import { JwtPayload } from "./interfaces/jwt-payload.interface";
 import { JwtToken } from "./interfaces/jwt-token.interface";
@@ -30,20 +33,23 @@ export class UsersService {
   ) {}
 
   async login({ email, password }: LoginUserDto): Promise<JwtToken> {
-    Logger.log(`User logged in: ${email}`);
-    const storedPasswd = (await this.usersRepository.findUser(email))?.password;
+    Logger.log(`User logging in: ${email}...`);
+    const storedPasswd = await this.usersRepository.getUserPassword(email);
     const login = await compare(password, storedPasswd ?? "");
     if (!login) {
       throw new UnauthorizedException(invalidLoginData);
     }
     const payload: JwtPayload = { email };
     const accessToken = this.jwtService.sign(payload);
+    Logger.log(`User logged in: ${email}.`);
     return { accessToken };
   }
 
   async register(registerDto: RegisterUserDto) {
-    Logger.log(`User created: ${registerDto.email}`);
-    return await this.usersRepository.createUser(registerDto);
+    Logger.log(`User creating: ${registerDto.email}...`);
+    const res = await this.usersRepository.createUser(registerDto);
+    Logger.log(`User created: ${registerDto.email}.`);
+    return res;
   }
 
   async changePwd(
@@ -78,10 +84,19 @@ export class UsersService {
   }
 
   async getUser(email: string) {
-    return await this.usersRepository.findUser(email);
+    const res = await this.usersRepository.findUser(email);
+    if (res == null) {
+      throw new NotFoundException("Felhasználó nem található.");
+    }
+    return res;
   }
 
-  async deleteUserByEmail(email: string) {
+  async deleteUserByEmail(email: string, user: User) {
+    if (!user.isAdmin && user.email !== email) {
+      throw new ForbiddenException(
+        "Csak adminok vagy a fiók tulajdonosa törölheti a fiókot.",
+      );
+    }
     return await this.usersRepository.deleteUserByEmail(email);
   }
 
@@ -93,6 +108,14 @@ export class UsersService {
 
   async requestToken(email: string) {
     await this.usersRepository.requestToken(email);
+  }
+
+  async search(search: SearchUserDto) {
+    const res = this.usersRepository.search(search);
+    if (res == null) {
+      throw new NotFoundException("Felhasználó nem található.");
+    }
+    return res;
   }
 
   /**
